Lot of things
Colision

+++ include/Vector.h
@@ -4,15 +4,15 @@
 
 struct Vector
 {
-    double x{0.f};
-    double y{0.f};
+    float x{0.f};
+    float y{0.f};
 
     void operator+=(Vector const &v);
     void operator-=(Vector const &v);
-    Vector operator *(double const &coef) const;
+    Vector operator *(float const &coef) const;
 
     //Rotation est un angle en °
-    static Vector GetDirection(double speed, double rotation);
+    static Vector GetDirection(float speed, float rotation);
 };
 
 #endif // VECTOR_H
